Add type_name()
[why]
Sometimes, especially when working with templates, one wants to know if
a type is really what one expected it to be.

[how]
There are mainly two techniques to debug the types:

At compile time one can get an output from the compiler by trying to
instantiate an incomplete type template. The resulting error message
contains a complete description of the type for which one tried to
instantiate the template.

At runtime one can resort to __PRETTY_FUNCTION__ that is present in
clang and GCC or __FUNCSIG__ of MSVC, and cut out all relevant parts.

To be able to manipulate string literals a static_string object is
useful, because that allows constexpr operations on the string literals.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>