statistics: Change return type of statistic properties
[why]
If we have a container of integers and want to calculate the mean value,
we usually want to get that value 'exactly' and not as integer.

[how]
Change the type used for the calculation and return value from the
container element_type to statistics_result_type which is at the moment
double.

[note]
Also changed the tests a bit, using better - hand calculated - target
values and ULP margins. Note that in case of float container elements
these numbers are already far off of the same number as double, so we
must check in ULP-of-float.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>