statistics: standard_deviation() returns sigma and mean
[why]
Often one needs both: the standard deviation value and the mean value of
some container.
As we have to calculate the mean value in any case (to be able to come up
with the standard deviation), we can also give that number back to the
user.
This avoids that the mean has to be calculated twice.

[how]
Instead of an out-parameter we decided after several discussion that a
specific struct would be a better solution to transport both values. A
cast operator helps in very simple use cases.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>