SlidingBuffer: Change logic operation from right-filled to left-filled
[why]
The operator[] had a strange inverted logic, with increasing indices
going to the 'left'. This could be unexpected.

[how]
If the filling of the buffer occurs on the left / beginning side, there
will be no 'inversion' and 'strangeness' on how the operator[] behaves.

If the logic is to push_front() instead of push_back() to fill elements
in, all other behavior is unchanged and easier to understand (maybe).

Accessing the underlying container is still 'unspecified order', so
there is no change there (it will now be some reverse order, but if it
is unspecified, this is not really a change).

[note]
Also hide push_back(), because users could try to use this wrongly, and
it is inherited from std::vector (if that underlying container is used).
This would go somehow unnoticed or hard to debug, so we effectively
remove that function and it's siblings.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>