statistics: Expose statistics_result_type
[why]
It might never be needed because usually you can infer the type from
the use cases; it might come in handy in some rare cases of template
work of a user of GUL.

There is no benefit in hiding it.

It's not documented in Doxygen if it's in an anonymous namespace, but
can be seen in Doxygen's description of functions as default template
parameter. Making it visible improves documentation.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>