statistics: Rework StandardDeviation{} object
[why]
1. The object also contains the mean value, so it should be named such
2. Data member names of (non-trivial) objects shall end with a underscore

[how]
1. Rename to StandardDeviationMean{}
2. Rename members, add getter functions

[note]
Also removed the possibility to instantiate the object with a data type
that is not floating point (e.g. int). The use case is rather arcane, and
because of changes in standard_deviation() some commits back it will not
be possible to create such an object from that function anymore. Maybe
it should have been removed here with that commit.

(In previous versions of standard_deviation() it was possible to get the
standard deviation of a container with integer elements as integer
value. This has been dropped recently as 'maybe not too usefull'.)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>