statistics: Remove latex code from Doxygen documentation
[why]
Generating the docs with Doxygen should be possible if all build time
dependencies are met. A 'works on my machine' with a specific setup is
not desirable.

It is more important to get the information to the user than to have a
nice formatting.

It is important to be able to build the docs to see if there are Doxygen
warnings and how they come out. Only then can authors contribute in a
meaningful way.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>