Remove static string
[why]
The static string (constexpr string) complicates the implementation and
only to get type_name() to be constexpr.

[how]
Replace the own string type with a gul::string_view. That also has
constexpr capabilities, but it is unclear how complete that is really
implemented.

Need to change the indices, because the full name of the function template
(that is used to cut out the type of the template parameter) is changed.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>