SlidingBuffer: Keep data in member instead of directly in object
[why]
There might be some drawbacks if one inherits from a standard container,
none of which appeared in this specific implementation. but one does not
know how the SlidingBuffer will evolve over time, and then it might
someday be problematic.

[how]
What we loose are all the different kind of constructors, that are
specific and different for the two underlying container variants. But
then, unually they should not be used anyhow.

Implement missing iterators.
Drop the at() member function.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>