SlidingBuffer: Implement simpler operator[]
[why]
The previous implementation of operator[] made too much guarantees, for
example it would never access memory outside the container. This is
usually only the case with at().

[how]
To remove a possible performance bottleneck we implement operator[], as
traditionally the case, without any checks and put the previous
implementation into the new function at().

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>