SlidingBuffer: Implement simple iterator for sliding access
[why]
All the iterator functions work directly with the underlying container
to get the best possible performance, but with out of order access to
the elements.
This is in contrast to operator[] and at() which access the elements in
the order they where pushed into the buffer.

[how]
This is a test implementation to get a simple forward-iterator that
walks through the elements in order.

For the time being the iterators can be accessed via begin_() and
end_(). We need to discuss how to proceed from here.

 - bidirectional iterators needed?
 - or random access iterators?
 - throw the direct access to the underlying container out
 - have a special function to get to that container
 - etc

This is not 100% finished, but as noted above I think we should discuss
first.

[note]
I tried to implement the iterator classically based on std::iterator,
but using the typedefs inside the iterator fails for some reasons. I
reckon too many stacked templates for the compiler. As std::iterator is
going to be deprecated anyhow we use the manual approach now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>