SlidingBuffer: Create a different implementation of iterators
[why]
Try to come up with a more efficient iterator design.

[how]
This new in-sequence iterator implementation works more closely to the
underlying buffer and removes as much as possible of the actual work to
do out of the increment and dereference functions.

The guarantees are less broad, though. But they usually are as limited
as with this implementation.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>