SlidingBuffer: Allow lvalue semantics
[why]
Previously the member function at() and operator[] returned an rvalue,
i.e. a constant reference to the actual element. It was not possible to
change the contents of the SlidingBuffer that has been pushed into it.

Although it might be dangerous for users to be able to change the
content of the SlidingBuffer it in itself is not problematic. It might
be a bit limiting in yet unknown use cases, so we remove this unusual
container limitation.

at() is added, because it has more guarantees than operator[] at a
slight performance cost.

[how]
Provide const and non-const member functions of at() and operator[].
The non-const members functions return lvalues now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>