SlidingBuffer: New implementation of SlidingBufferIterator
[why]
After some experimentation this is the new iterator version. It is more
complete than the previous ones (const and non-const iterators), and it
has the stronger guarantees then the last iterator.

[how]
To get const and non-const iterators without much duplicated code a
template is used, where a template parameter decides which version is
instantiated.

Documentation is completed and iterator invalidation described.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>