Revert "SlidingBuffer: Extract SlidingBufferIterator out of main class"
[why]
I find the decision very tough. But I reckon this can be interpreted as
a case of Core Guidelines C.5:

  Place helper functions in the same namespace as the class they support

  A helper function is a function (usually supplied by the writer of a
  class) that does not need direct access to the representation of the
  class, yet is seen as part of the useful interface to the class.
  Placing them in the same namespace as the class makes their relationship
  to the class obvious and allows them to be found by argument dependent
  lookup.

I think this also holds for a helper (sub-) class.

[how]
This reverts commit 247929d3c6811cdb5f7dc4b9daaf5bb8b13d26b8.

[note]
This and the previous commit have been left in the repository for
documentation purposes.

Minor changes in documentation compared to reverted commit.