SlidingBuffer: Simplify iterator comparison op
[why]
I thought a lot about this change. Before the change one could compare
iterators of different containers and they would never be equal. But it
is an additional computational burden.

The C++ standard requires that iterators compared are from the same
container, or the result is undefined.

This relates to C++ standard library issue 466, resolved with N3066.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>