SlidingBuffer: Add SlidingBufferExposed{}
[why]
For use cases where the iterator interface is just used to process 'all
data' and the order of elements is not important, where the performance
of iterating through the buffer is important: we reinvent the disjunct
interator and index interfaces.

This is likely to be only important on deep buffers, where iterating
itself takes most of the time.

[how]
Inherit the SlidingBuffer and just replace the iterator interface.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>