SlidingBuffer: Add copying/moving tests
[why]
We want to be sure that the data structure is indeed (deep) copied, if
we write code for a copy.

And we want to be sure it is moved if we say so.

This can easily break if changes to the SlidingBuffer is made.
(For some definitions of 'easy'.)

[how]
Copy: Modify deep member and check in copied from buffer.
Move: Compare addresses of moved from and to deep member.

[note]
For the floating point values we take integer numbers, as these are
guaranteed to compare equal :->

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>