SlidingBuffer: Make iterators copyable
[why]
We all like to copy iterators, do we not?
There is no reason to not allow this.

[how]
I struggle to write a move test, with no access to the members, but
probably this is a non-issue, because they are lightweight anyhow (who
would want to move and could not live with a copy).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>