SlidingBuffer: Rename BufferSize nontype template parameter
[why]
As the buffer size is not a type we fall back to the style guide that
'variables' use snail_case. This is in fact often done in the
literature, but not always (that's style).

Notable references that use CamelCase for nontype template parameters is
the Core Guidelines examples and the C++ standard library.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>