type_name: Rewrite from scratch
[why]
The previous implementation failed on Darwin's clang.

[how]
This time we rely on the constexpr operations of string_view. This makes
a more forgiving parsing of the string.

[note]
Noticed that __PRETTY_FUNCTION__ is **not** constexpr in some versions
of GCC. In that case the whole function is not constexpr.

MSVC also has problems with constexpr in some instances.

Tested on Clang 3.8, Clang 6, GCC 5.4, GCC 7.4, Apple Clang 8 and MSCV 19.
Tested on Ubuntu and Darwin targets.

[note]
std::strlen() is not constexpr; while strlen() is in some compilers.
We use sizeof()-1 instead.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>