type_name: Fix constexpr-ness
[why]
After again checking the constexpr-ness of the function I found out that
some standard-library functions are NOT constexpr despite being
specified as such.
So instantiation in a constexpr context failed.

[how]
Create constexpr versions of the relevant functions (in an anonymous
namespace) and use them in type_name().

[note]
Also add missing include.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>