SlidingBuffer: operator<< dumps now a user's view
[why]
The dumping to a ostream is a tool for users of the library to have a
quick insight into the buffer's contents. They should not need to know
any implementation details to interpret the data.

[how]
Just dump the elements in order.

[note]
We loose a lot of tests that check the implementation details, as there
is now no way anymore to inspect the real contents of the underlying
container.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>