SlidingBuffer: Re-add tests removed with previous commit
[why]
It seems beneficial to really check what is going on inside the
underlying buffer of the SlidingBuffer. With the last commit we lost the
possibility for any inspection.

[how]
Implement a debug version of the SlidingBuffer that just adds in
inspection method with output identical to the removed operator<<
functionality.

This can we use to have SlidingBuffer that are able to introspect.

Use the inspection to re-enable all the tests we had before the previous
commit.

So now:
 - No internals inspection possible with normal SlidingBuffer
 - Define thin 'overlay' object (in test setup) to re-enable inspection
 - Inspect internals as before the previous commit

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>