finalizer: Add second finally call with move semantic
[Why]
When we can move the closure we should do it. It could have fetched
a lot of expensive to copy or even move only objects.

The original finally function took a const ref, which was passed
to the only actual constructor. The new version is using the move
constructor and is forwarding it, as done in the GSL version.

[how]
Just move it, if possible. If not, copy. Then move into final_action{}.

Previously we took by reference and then always copied in the
final_action ctor. Also add a test case for the new call.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>