deb: Limit ARM targets to sequential packeting
[why]
The ARM target(s), i.e. Raspberry, has a resource/memory problem when
doing things in parallel while building and testing.

The building via 'make' (*) is already thread limited in the Makefile,
but building via 'debuild' or 'makeDdeb' is not, because that uses
the debian/rules file.

(*) on ARM targets

[how]
Create an appropriate debian/rules file, that limits the ninja threads
to one if the build system is an arm.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>