make: Use different prefix on Darwin
[why]
On Darwin the library has to be installed in '/local/lib' instead of
'/export/doocs/lib'.

[how]
When we detect the host OS to det the correct cpp-args we can also set
the correct prefix directory.

The code now acutally uses the variable PREFIX, previously it was set
but not used :->

[note]
This is a fix as discussed with Tim today. This is untested, as I have
not the means to really test this to the end.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>