finalizer: Correct universal reference behavior
[why]
Although the previous version works correctly (in most? cases), it
expresses poorly what is really happening here. Change the formulation
so that the real intent is more clear.

[how]
The 'template <typename T> finally(T&& t)' is a universal forwarding
template, and as such almost always matches better than our other
template specialisations of finally. This means this function would be
used/called in most situations, even if T is an lvalue etc - our other
finally() versions are not even considered in overload resolution.

This differs from a non-template function like 'finally(SomeType&& t)'
which would not bind to lvalues references.

In non-templated functions we usually have two functions like:
  1: func(Type const& t)
  2: func(Type&& t)

With templates this is not possible, because when we use
  3: template <typename Type> func(Type const& t)
  4: template <typename Type> func(Type&& t)

calls that do not exactly match 3 (for example because the call
parameter is not const in itself) will end up in 4. This is because 3
needs an implicit conversion (non-const to const), while 4 always
matches perfectly (thus: perfect forwarding).

But what we want to setup here is the behaviour we would expect from 1
and 2, just that the Type is not known.

What we can do is SFINAE out formulation 4 when the parameter is not an
rvalue reference. lvalues will then be forced to use 3 (after possible
implicit conversion).

I think this is an important point to make. Technically the previous
statement was correct with it's movability check, but this conveys the
intent better imho.

[note]
The problem is described in many places, I'd like to reference here
Scott Meyers article in 'overload'.
http://scottmeyers.blogspot.com/2012/10/new-article-universal-references-in-c11.html
And a talk Scott presented at C++ and Beyond 2012:
http://scottmeyers.blogspot.com/2012/10/video-for-universal-references-in-c11.html

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>