lint: Allow single-line braces-less if cases
[why]
A new rule warns on cases where a if/for/do/while statement is followed
by a scope-less (without braces) statement. For example:

if (test) do_stuff();

if (test)
    do_stuff();

if (test)
    x = do_stuff()
        * and_do_more();

The expected (more supposedly more reasable versions) are:

if (test) { do_stuff(); }

if (test) {
    do_stuff();
}

if (test) {
    x = do_stuff()
        * and_do_more();
}

[how]
Although we generally do not use the first case anywhere in the code
(all in one line), the second form is very widespread in GUL. The third
form is really a bit hard to read maybe.

From this we would set the line limit to '2', catching the third case,
but not the second or first.

But unfortunately if we have an else branch, the line with the else
keyword also counts, so we need to set it to '3' to allow

if (test)
    do_stuff();
else {
    do_other_stuff();
}

But this means we do not warn on case 3 from the top anymore.

For me this breaks the usability of the whole check. But anyhow,
we do not want it to give constantly false hits.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>