cat: Take parameters as const&
[why]
There is no need to copy the ConvertingStringView objects if we just
want to read access their contents (to push it into a std::string).

Usually the ConvertingStringView are temporaries that will be moved in,
so in the ordinary cases this is not a problem (does not happen by
language rules and can (will) be optimized away).

[note]
The initializer_list version will be handled in the next commit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>