cat: Simplify formulation of variadic template
[why]
To improve the readability (ok, this is relative) and for a simpler
construction of the initializer_list.

[how]
The variadic template is now only triggered if we have enough arguments
anyhow, so the need for dummy parameters and their inclusion in the list
vanishes.

[note]
This still warns 'do not implicitly decay an array into a pointer', but
I don't know how to fix that.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>