statistics: Remove unwanted copying by remove_outliers()
[why]
The remove_outliers() function works 'in place' (i.e. on the container
that is passed to it). For conveniance (ability to chain operations) it
also returns a reference to the processed (== the original) container.

Well, that was the design idea. As clang-8 points out this is not the
case with the previous code. It rather copies the modified container and
return that 8-(

[how]
Add forgotten ref operator...

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>