meson: Let Doxygen target depend on correct files
[why]
Previously the data/docs target was only evaluated once, because it just
had itself as dependency.

This means the doxygenerated pages can not be easily regenerated after
changes.

[how]
All implementation (i.e. .cc) and public header files are added to the
dependencies. Private headers (that we do not have) are exempt.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>