finalizer: Use gul::toc() in timing example
[why]
This also shows off how easy it is to time something with the
gul/time_util.

[how]
Use capture with initializer in the lambda expression to avoid an
additional 'start' variable. This also might be seen as an example
how things can be done with closures.

[note]
So start is 'auto' without even saying 'auto' :->
Soeren will hate me.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>