finalizer: Fix move assignment
[why]
Move assignment is only possible for non-closure actions (apart from
self assignment, because each closure has its own distinct type). So it
is only viable for function pointer actions. But that was not tested.

I noticed that the code looked wrong, with this->member_(val) used to
construct(?)/assign member_ with val. I reckon this is a copy&paste
mistake from the move ctor definition. This is the reason why I prefer
to use brace-initialisation, because then it would have been obvious.
No new object created there, so {} looks funny. Now the compiler
'thought' that we want to call the action, with the new action as
argument. Which of course breaks.

[how]
Add test that fires. Remove wrong/meaningless test.
Use brace initialisation.
Use move assignment for members.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>