finalizer: Simplify finally() by dropping one overload
[why]
There were two overloads for finally:
 1. One for lvalue references that need to be decayed first
 2. One for rvalue references, rvalues and glvalues (lvalues and xvalues)

I had the decay in a different overload, because I did not want to have
the lambda rvalues turned into a reference; that would create possible
lifetime problems.

But a hard look on std::is_function showed that lambdas are explicitely NOT
functions in that sense. So we can safely apply std::decay on the rvalue
lambdas without the decay happening. They will have cv removed
(std::remove_cv), but I can not come with any scenario where this has an
impact.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>