Slidingbuffer: Circumvent Doxygen warning
[why]
Doxygen does not resolve type aliases (in most cases). To generate a ref
to a function signature the used aliased type have to be used (most
often).

[how]
Either we use the not so nice alias name, or we drop the reference. At
least this is what I think is possible with Doxygen.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>