cat: Prevent initializer list usage with few arguments
[why]
When only a few arguments to cat() are supplied we want to use the
special implementations for that number of arguments. But unfortunately
the enable_if that prevents the better matching template to be
considered did not work.

[how]
Write code that really is an error and triggers SFINAE when the template
is not to be chosen.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>