SlidingBuffer: Rename template parameter buffer_size
[why]
The template parameter 'buffer_size' describes the size of the
underlying container and this is where the name originally came from.
When viewed from a user's perspective this is not expected and would
rather be called 'capacity' of the SlidingBuffer.

[how]
The parameter is renamed to fixed_capacity.

The reason for the fixed_ prefix stems from the fact that we have two
SlidingBuffer variants:
- with a variable capacity using on std::vector for storage
- with a fixed capacity using on std::array for storage

Not only describes the template parameter the 'initial' capacity of the
underlying container and the capacity of the SlidingBuffer, it also
says if the SlidingBuffer is of variable capacity: Variable when
argument is zero.

So the new renamed parameter can be interpreted in two ways, which is
aided by the new name:

As bool: !!fixed_capacity shows if SlidingBuffer has a fixed capacity
As int:  fixed_capacity's value is the capacity - that is fixed

Reported-by: Lars Froehlich <lars.froehlich@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>