meson: Allow configuration without working git
[why]
Git is no build time dependency and absence or failures should be
handled gracefully and result in a un-revisioned library.

This is not the case if git fails sometimes and sometimes not.

[how]
Handle 'unknown' versions in a meaningful manner, as it has been
implemented in all other places. This one place has been overlooked.

Reported-by: Tobias Tempel <tobias.tempel@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>