meson: Use API version if git does not give version numbers
[why]
With the previous commit we (re)added the possibility to build the
library without having git (or a incompatible version of git).

The generated library object had the dummy version '0.0.1', which might
be strange for some.

[how]
We take the API version as library version instead of '0.0.1'. This is
almost the same we would get with a working git, just the patchlevel is
missing.
An informative message is also printed.

[note]
We have two concurrent version schemes: The API version (that is
hardcoded in the main meson.build) and the more detailed git-version.

It is always checked, that the git-version and the API version are the
same in major and minor. The defined API version does not have a
patchlevel at all. The patchlevel is determined with git, i.e. the
number of patches since the first occurence of the new API version.

With this solution we loose the patchlevel, but because it is only
existing in git, and closely related to git commits, this can not be
helped.

So the API version has to be set manually/explicitly in the main
meson.build file. There are some (not specified) reasons, also discussed
on stash, that make this approach desirable.

The git database is used to fill in the small bits: It provides an
automatic patchlevel and facilitates autogeneration of the package
changelog.

So git is not a build time dependency, which I would not like.
But git is a packaging dependency, together with some more tools not
explicitely stated in the debian/control file.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>