SlidingBuffer: Add push_back()
[why]
There is no reason why the sliding buffer should be limited
to sliding in one direction. By adding push_back(), the class
can be made more versatile without sacrificing compatibility
with existing use cases. Users are free to select the direction
that is most suited to their application and might even mix
push_front() and push_back().

[how]
The original data ordering in the underlying buffer was opposite
to the logical ordering exposed by SlidingBuffer. For adding
push_back(), I changed the order to conform with the underlying
container. This means that push_front() now appends on the left
(at lower indices) whereas push_back() appends on the right (at
higher indices). This broke many tests that checked the internal
memory layout of the buffer; they are now commented out.