SlidingBufferIterator: Remove virtual destructor
[why]
An iterator should never be used polymorphically,
therefore there is no need for a virtual destructor.
Additionally, the vtable generated because of it
increases the memory footprint of an iterator instance
dramatically.

[how]
Remove the virtual destructor definition. This also
makes the other defaulted constructors redundant
(they are defaulted ... well, by default), so they
are also removed.