SlidingBufferExposed: Fix begin() and end()
[why]
They did not work as expected with push_front() or with
mixed usage of push_front() and push_back().

[how]
begin() and end() now return iterators to the first or past the
last element of the occupied range if it is contiguous. Otherwise,
they just return begin() and end() of the underlying container.