tests: SlidingBuffer: Adapt debug-versions to new bidirectionality
[why]
The recent changes added a new data member, the beginning index of the
data. The position of that index was not visible in the old debug
versions.

[how]
Show the beginning index with a 'b', and the end index with an 'e'.

If both, beginning and end point to the same element, show its index
with a '*'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>