SlidingBuffer: Expand begin()/end() guarantees
[why]
With the bidirectional SlidingBufferExposed we get suboptimal begin()
and end() values for un-full buffers (i.e. just spanning the total
underlying buffer) in some cases:
  When the push_front() and push_back() are both used on the same buffer
  When push_front() and resize() is both used on the same buffer

The first case (mixing front and back insertion) can be discarded
as rather unusual case, where users would have to use the un-exposed
SlidingBuffer.

But the later case is not too unusual and should not make the exposed
version unusable.

[how]
Detect if only push_front()s have been used, and if that was the case do
not re-align the data in the underlaying container in the push_back()'s
style, but rather keep it right/end aligned.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>