SlidingBuffer: Reimplement 'expand begin()/end() guarantees'
[why]
Why-reimplement:

It has been implemented in the normal SlidingBuffer, where it was
not needed to fulfill the begin() and end() guarantees. To
disentangle the code requirements of SlidingBufferExposed and
SlidingBuffer the implementation should be shifted into the
affected object, i.e. SlidingBufferExposed.

Why-implement:

The iterator interface of SlidingBufferExposed (i.e. begin(), ...)
had (and has) this specification:

> The iterator accesses the underlying container in its native order. If the elements
> are stored in contiguous fashion (if only push_back() or only push_front() were
> used to add them), begin() returns an iterator pointing to the first occupied
> element. If the elements are non-contiguous (which can happen if push_front() and
> push_back() were mixed and the buffer is not yet full), begin() simply points to
> the start of the container and the range [begin(), end] may enclose
> default-constructed elements.

This guarantee, that - if you refrain from mixing push_back() and
push_front() - the iterators will only access valid elements is quite
useful and is one key aspects that make the exposed version fast but
correct.

But there is a problem when push_front() is exclusively used (i.e. no
push_back()), but the buffer size has to change while running. Due to
the way the resizing is implemented the iterators will behave as if the
elements are non-contiguous. This breaks basic functionality of the
exposed buffer.

[how]
Partially revert (just keep test cases)
  b936a84 SlidingBuffer: Expand begin()/end() guarantees

After that the actual resizing functionality is moved from resize() to a
protected member function. Now resize() and reserve() of SlidingBufferExposed
are introduced that shadow the SlidingBuffer's members.

The resizing in the exposed version checks for the corner case and
handles it suitably. If the corner case is not present it just calls the
new protected member function to resize in the 'normal' way.

Shadowing is selected over virtuality because polymorphism is not expected here.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>