SlidingBuffer: Add systematic resize tests
[why]
To see what is going on, and if that is what we want.

[how]
Just do systematic stuff and dump it in a nice way. Use the meson log
file 'meson-logs/testlog.txt' to see the output after performing test.

Afterwards we can either revert this commit or turn the stuff into real
tests.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>