SlidingBuffer: Avoid trailing whitespace in debug dump
[why]
The debug dump output had trailing blancs which are not needed and can
be avoided to facilitate easier automatic processing of it.

[how]
Check ahead of actual dumped element is the last one, and avoid the
blancs if it is.
The code is not very optimized, but this is for debugging.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>