SlidingBuffer: Allow shrinking in different directions
[why]
Previously, if a SlidingBuffer's size has been decreased the front
elements have been kept, so that
buf_old[n] == buf_new[n] for all n within the new size.

If the buffer has been filled with push_back() this means that the most
recent data is dropped which might be unexpected. If filling is
continued there will be a 'time gap' in the contained data.

For this use case it would be good to keep the back elements, so that
buf_old[old_size - n] == buf_new[new_size - n] for all n within new_size.

[how]
Add a parameter which part of the data is to be kept on shrinking the
buffer.

The decicion can not be automated if we want to allow mixed use of
push_front() and push_back() on the same buffer, so the user has to decide
what is most appropriate.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>