SlidingBuffer: Fix SlidingBufferExposed shrinking
[why]
In some cases the formula used to determine the shift needed prior to
shrinking failed.

The different (in comparison with SlidingBuffer<>) approach is needed,
because the data has to be kept on the right hand side of the underlying
container to avoid gaps with uninitialized elements between the begin
and end iterators.

[how]
The old formula did not properly expect the need to drop data when the
unfilled buffer becomes filled by the resize.

Now we rotate only as far as the beginning of the filled data is,
turning the begin to underlying index 0. If we would rotate more we
would later loose recent data.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>