SlidingBuffer: Fix iterator documentation
[why]
The iterator invalidation description was partially outdated, partially
wrong, and partially misleading.

[how]
I hope the changes help to reduce the misleading-ness and remove the
outdated and wrong parts.

Moved the description of invalidation into the description of the
iterators themselfes (from the buffer).

Also corrected the invalidation description of the exposed version's
iterators.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>