Add copy&move constructors to struct HexdumpParameterForward
[why]
The struct stores iterators to its internal container. These iterators can be
invalidated if the struct is moved or copied, so we need custom assignment
operators and copy/move constructors to update the iterators.

[how]
Add the various constructors and assignment operators and have them call a
private member function regenerate_iterators() which stores the new begin() and
end() if cont_ is a container.