tests:SlidingBuffer: Add static_cast to test for at(-1) to suppress VS2019 warning
[why]
The test intends to assert that accidentally passing a negative value to at() throws an exception. VS2019 correctly issues a warning that a signed integer is passed into an unsigned argument.

[how]
Use a static_cast to ensure the compiler that we know what we are doing. Also, add some comments to remind our future selves why this is necessary.