num_util: Broaden types processed by within_orders()
[why]
The type to be compared has been limited to floating point types.
Contrary to that the example given in the Doxygen doc uses integers.

There is no real reason to limit the type.

[note]
I think the limit came from the previous, needlessly complicated
implementation.

Stumbled about this when working with generic code that is sometimes int
and sometimes float, which could not instantiate with int :-(

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>