escape: Optimize runtime of unescape() for shorter strings
[why]
For consistency I tried to use a preallocated string output buffer (like
escape() uses) instead of the stringstream. I thought that the
preallocation of the output buffer might help to get faster, while
sometimes allocating slightly more than necessary (for each escape at
least the '\\' is dropped.

While the difference for long strings is not much (speedup of the new
code decreases slowly to something just above 1.0x as the string length
increases), it is always faster, up to more than 2.0x for very short
strings.

[how]
Use a preallocated string as output buffer. Use the inputs size as the
size of the buffer. The output string will always be at most this long.
If escaped chars are present the output will be shorter and we have
allocated too much. This is because

   \\n   ->   \n    1 character size decrease
   \\x40 ->   @     3 character size decrease

So in principle if there are lots and lots of escape sequences present
the old code might be better (memory wise, but not speed wise).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>