tests:SlidingBuffer: Make dump configurable
[Why]
Instead of removing the debug messages, we move them into a higher debug
level. This way they are not visible during normal runs, but be still
around if we need them.

[How]
Add a new argument to do_a_dump indicating messages should be printed or
not. And adopt the calls accordingly.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>
--
v2: Add missing variable on second test section