to_number: Make to_number() noexcept and return an optional<ValueType>; improve floating-point conversion
[why]
This is an alternative implementation to the previous one that
relied on exceptions to communicate conversion errors. When parsing
text, to_number() might potentially be used on a huge number of strings
that are not valid numbers. According to the guideline "use exceptions
only for exceptional cases", it could be argued that the implementation
with optional<> should be preferred. In any case, it is expected to be
faster for invalid input.

Additionally, the commit improves the implementation of the
floating-point conversion and adds more test cases.