Add GUL_EXPORT macro to handle DLL symbol export and import on Windows
[why]
On Windows, symbols for dynamic link libraries need to be exported
explicitly when the library is built. They also need to be imported
explicitly when the DLL is consumed. If a static version of the
library is consumed, there may be neither an import nor an export
statement.

[how]
Add an include file "internal.h" that defines the macro GUL_EXPORT.
On Visual C++, GUL_EXPORT expands to one of the following:
- "__declspec(dllexport)" if a DLL is being built (as indicated by
  a macro set by Meson)
- "__declspec(dllimport)" if a DLL is being consumed (this is
  treated as the default case)
- "" (an empty string) if linking against a static version of the
  library. This must be indicated explicitly by defining the macro
  GUL_USING_STATIC_LIB_OR_OBJECTS.
On compilers other than Visual C++, the macro is defined but empty.